backend/ingestion/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
cohere>=4.5.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0